perf: Convert landing page to server component with client interactive islands
